refactor: move modals into modals folder
