+++ src/app/dashboard/company/club/new/page.tsx
@@ -1,8 +1,8 @@
 import ApplicationOverview from "@/components/dashboard/company/club/molecules/ApplicationOverview";
 import ApplicationView from "@/components/dashboard/company/club/templates/ApplicationView";
-import ApplicantProfileModal from "@/components/dashboard/company/club/molecules/ApplicantProfileModal";
-import RejectApplicationModal from "@/components/dashboard/company/club/molecules/RejectApplicationModal";
-import RevertRejectionModal from "@/components/dashboard/company/club/molecules/RevertRejectionModal";
+import ApplicantProfileModal from "@/components/dashboard/company/club/modals/ApplicantProfileModal";
+import RejectApplicationModal from "@/components/dashboard/company/club/modals/RejectApplicationModal";
+import RevertRejectionModal from "@/components/dashboard/company/club/modals/RevertRejectionModal";
 
 export default function NewApplicationPage() {
   return (
