+++ src/app/dashboard/company/club/page.tsx
@@ -1,6 +1,6 @@
 import ClubOverview from "@/components/dashboard/company/club/molecules/ClubOverview";
 import ClubView from "@/components/dashboard/company/club/templates/ClubView";
-import ApplicantProfileModal from "@/components/dashboard/company/club/molecules/ApplicantProfileModal";
+import ApplicantProfileModal from "@/components/dashboard/company/club/modals/ApplicantProfileModal";
 
 export default function Page() {
   return (
